move auth wrapper to layout directory enable deleting meetings in sandbox

+++ src/utils/api/events.ts
@@ -1,6 +1,7 @@
 import { api } from './apiConfig'
 import { buildSandboxEvent } from 'utils/helpers'
 import { EventInfo } from 'interfaces'
+import { isSandboxId } from 'utils/helpers/taskHelpers'
 
 export const createEvent = async (calendarId: string, eventInfo: EventInfo) => {
   try {
@@ -36,6 +37,10 @@ export const updateEvent = async (calendarId: string, eventId, eventInfo) => {
 
 export const deleteEvent = async (calendarId: string, eventId) => {
   try {
+    if (isSandboxId(calendarId)) {
+      return
+    }
+
     const res = await api.delete(
       `/calendar/${calendarId}/fetchEvent/${eventId}`
     )
